+++ test_spuce/plot_fft.cpp
@@ -7,12 +7,12 @@ using namespace spuce;
 #include <iostream>
 
 void plot_fft(std::vector<double>& data) {
-  const long N = data.size();
+  const size_t N = data.size();
   Dictionary d;
   d.add("lw", 3.1);
 
   std::unique_ptr<std::complex<double>[]> z(new std::complex<double>[N]);
-  for (unsigned int i = 0; i < N; i++) z[i] = data[i];
+  for (size_t i = 0; i < N; i++) z[i] = data[i];
 
   cfft F(N);
   F.fft(z.get());
@@ -20,7 +20,7 @@ void plot_fft(std::vector<double>& data) {
   std::vector<double> fft(N/2);
   std::vector<double> x(N/2);
 
-  for (int i = 0; i < N/2; i++) {
+  for (size_t i = 0; i < N/2; i++) {
     double val = norm((double)N * z[i]);
     if (val < 1e-6) val = 1e-6;
     fft[i] = 10.0 * log10(val);
@@ -33,12 +33,12 @@ void plot_fft(std::vector<double>& data) {
   plt.show();
 }
 void plot_fft(std::vector<std::complex<double>>& data) {
-  const long N = data.size();
+  const size_t N = data.size();
   Dictionary d;
   d.add("lw", 3.1);
 
   std::unique_ptr<std::complex<double>[]> z(new std::complex<double>[N]);
-  for (unsigned int i = 0; i < N; i++) z[i] = data[i];
+  for (size_t i = 0; i < N; i++) z[i] = data[i];
 
   cfft F(N);
   F.fft(z.get());
@@ -46,7 +46,7 @@ void plot_fft(std::vector<std::complex<double>>& data) {
   std::vector<double> fft(N);
   std::vector<double> x(N);
 
-  for (int i = 0; i < N; i++) {
+  for (size_t i = 0; i < N; i++) {
     double val = norm((double)N * z[i]);
     if (val < 1e-6) val = 1e-6;
     fft[i] = 10.0 * log10(val);
@@ -71,19 +71,19 @@ void plot_data(std::vector<double>& data) {
 
 void compare_fft(std::vector<double>& data1, std::vector<double>& data2)
 {
-  const long N=data1.size();
+  const size_t N=data1.size();
 	Dictionary d;
 	d.add("lw", 3.1);
 
 	std::unique_ptr<std::complex<double> []> z(new std::complex<double>[N]);
-	for (unsigned int i = 0; i < N; i++) z[i] = data1[i];
+	for (size_t i = 0; i < N; i++) z[i] = data1[i];
 	
   spuce::cfft F(N);
 	F.fft(z.get());
 
 	std::vector<double> fft(N/2);
 
-	for (int i=0;i<N/2;i++) {
+	for (size_t i=0;i<N/2;i++) {
 		double val = norm(N*z[i]);
 		if (val < 1e-6) val = 1e-6;
 		fft[i] = 10.0*log10(val);
@@ -93,10 +93,10 @@ void compare_fft(std::vector<double>& data1, std::vector<double>& data2)
 	plt.init();
 	plt.plot_data(fft, &d);
 
-	for (unsigned int i = 0; i < N; i++) z[i] = data2[i];
+	for (size_t i = 0; i < N; i++) z[i] = data2[i];
 	F.fft(z.get());
 	
-	for (int i=0;i<N/2;i++) {
+	for (size_t i=0;i<N/2;i++) {
 		double val = norm(N*z[i]);
 		if (val < 1e-6) val = 1e-6;
 		fft[i] = 10.0*log10(val);
